+++ src/mock/films-mock.js
@@ -1,3 +1,11 @@
+import {
+  getRandomNumber,
+  getRandomIntegerNumber,
+  getRandomDate,
+  getRandomArrayElements,
+  getRandomArrayItem
+} from '../util';
+
 const MONTH_NAMES = [
   `January`,
   `Fabruary`,
@@ -185,50 +193,6 @@ const TEXT = [
 ];
 
 const FILM_AGES = [`0+`, `6+`, `12+`, `16+`, `18+`];
-
-const getRandomNumber = (a = 1, b = 0) => {
-  const lower = Math.min(a, b);
-  const upper = Math.max(a, b);
-
-  return lower + Math.random() * (upper - lower);
-};
-
-const getRandomIntegerNumber = (a = 1, b = 0) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-
-  return Math.floor(lower + Math.random() * (upper - lower + 1));
-};
-
-const getRandomArrayItem = (array) => {
-  const randomIndex = getRandomIntegerNumber(0, array.length - 1);
-
-  return array[randomIndex];
-};
-
-const shuffleArray = (array) => {
-  for (let i = array.length - 1; i > 0; i--) {
-    let j = getRandomIntegerNumber(0, array.length);
-    let temp = array[i];
-    array[i] = array[j];
-    array[j] = temp;
-  }
-
-  return array;
-};
-
-const getRandomArrayElements = (array, min, max) => {
-  const randomMax = getRandomIntegerNumber(min, max);
-  const newArray = array.slice();
-
-  shuffleArray(newArray);
-
-  return newArray.slice(0, randomMax);
-};
-
-function getRandomDate(start, end) {
-  return new Date(start.getTime() + Math.random() * (end.getTime() - start.getTime()));
-}
 
 const generateFilmComment = () => {
   return {
